automation/requirements.txt
google-cloud-firestore
google-cloud-secret-manager
requests
pandas
cryptography